# core-common-document-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-document-mgmt
  description: Enterprise Content Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: https://api.getfirefly.io/ecm
    description: Production Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/documents/{id}:
    get:
      tags:
        - Document Controller
      summary: Get document by ID
      description: Returns a document by its ID
      operationId: getDocumentById
      parameters:
        - name: id
          in: path
          description: ID of the document to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved document
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
    put:
      tags:
        - Document Controller
      summary: Update an existing document
      description: Updates an existing document by its ID
      operationId: updateDocument
      parameters:
        - name: id
          in: path
          description: ID of the document to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentDTO'
        required: true
      responses:
        '200':
          description: Document updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid document data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
    delete:
      tags:
        - Document Controller
      summary: Delete a document
      description: Deletes a document by its ID
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          description: ID of the document to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found
...
  /api/v1/documents/{documentId}/permissions/{id}:
    get:
      tags:
        - Document Permission Controller
      summary: Get specific permission
      description: Returns a specific permission for a document
      operationId: getDocumentPermission
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          description: ID of the permission to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved document permission
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentPermissionDTO'
        '404':
          description: Document permission not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentPermissionDTO'
    put:
      tags:
        - Document Permission Controller
      summary: Update a document permission
      description: Updates an existing permission for a document
      operationId: updateDocumentPermission
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          description: ID of the permission to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentPermissionDTO'
        required: true
      responses:
        '200':
          description: Document permission updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentPermissionDTO'
        '400':
          description: Invalid document permission data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentPermissionDTO'
        '404':
          description: Document permission not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentPermissionDTO'
    delete:
      tags:
        - Document Permission Controller
      summary: Remove a permission from a document
      description: Deletes a permission from a document
      operationId: deleteDocumentPermission
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          description: ID of the permission to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Document permission deleted successfully
        '404':
          description: Document permission not found
...
components:
  schemas:
...
    DocumentDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the document
          format: uuid
          readOnly: true
        name:
          type: string
          description: Name of the document
        description:
          type: string
          description: Description of the document
        fileName:
          type: string
          description: Original file name of the document
        fileExtension:
          type: string
          description: File extension of the document
        mimeType:
          type: string
          description: MIME type of the document
        fileSize:
          type: integer
          description: Size of the document file in bytes
          format: int64
        documentType:
          type: string
          description: Type of the document
          enum:
            - DOCUMENT
            - IMAGE
            - PDF
            - CONTRACT
            - INVOICE
            - RECEIPT
            - FORM
            - REPORT
            - EMAIL
            - AUDIO
            - VIDEO
            - ARCHIVE
            - OTHER
        documentStatus:
          type: string
          description: Current status of the document
          enum:
            - DRAFT
            - UNDER_REVIEW
            - APPROVED
            - REJECTED
            - PUBLISHED
            - ARCHIVED
            - MARKED_FOR_DELETION
            - DELETED
            - LOCKED
            - EXPIRED
        storageType:
          type: string
          description: Storage type of the document
          enum:
            - LOCAL_FILESYSTEM
            - DATABASE
            - S3
            - AZURE_BLOB
            - GOOGLE_CLOUD_STORAGE
            - CDN
            - DISTRIBUTED_FS
            - EXTERNAL_REFERENCE
        storagePath:
          type: string
          description: Path where the document is stored
        securityLevel:
          type: string
          description: Security level of the document
          enum:
            - PUBLIC
            - INTERNAL
            - CONFIDENTIAL
            - RESTRICTED
            - SECRET
            - TOP_SECRET
        folderId:
          type: string
          description: ID of the folder containing the document
          format: uuid
        isEncrypted:
          type: boolean
          description: Indicates if the document is encrypted
        isIndexed:
          type: boolean
          description: Indicates if the document is indexed for search
        isLocked:
          type: boolean
          description: Indicates if the document is locked for editing
        lockedBy:
          type: string
          description: User who locked the document
        lockedUntil:
          type: string
          description: Date and time until the document is locked
          format: date-time
        expirationDate:
          type: string
          description: Date and time when the document expires
          format: date-time
        retentionDate:
          type: string
          description: Date and time until the document should be retained
          format: date-time
        tenantId:
          type: string
          description: Tenant ID for multi-tenancy support
        createdAt:
          type: string
          description: Date and time when the document was created
          format: date-time
        createdBy:
          type: string
          description: User who created the document
        updatedAt:
          type: string
          description: Date and time when the document was last updated
          format: date-time
        updatedBy:
          type: string
          description: User who last updated the document
        version:
          type: integer
          description: Version number for optimistic locking
          format: int64
        checksum:
          type: string
          description: Checksum of the document file for integrity verification
      description: Document data transfer object
...
    DocumentPermissionDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the permission
          format: uuid
        documentId:
          type: string
          description: ID of the document this permission applies to
          format: uuid
        partyId:
          type: string
          description: ID of the party this permission applies to
          format: uuid
        permissionType:
          type: string
          description: Type of permission
          enum:
            - READ
            - WRITE
            - DELETE
            - SHARE
            - EXECUTE
            - CREATE
            - MOVE
            - COPY
            - VIEW_METADATA
            - MODIFY_METADATA
            - VIEW_VERSIONS
            - CREATE_VERSION
            - VIEW_AUDIT
            - MANAGE_PERMISSIONS
            - CHECKOUT
            - CHECKIN
            - SIGN
            - SEND_FOR_SIGNATURE
            - ADMIN
        isGranted:
          type: boolean
          description: Indicates if the permission is granted or denied
        expirationDate:
          type: string
          description: Date and time when the permission expires
          format: date-time
        tenantId:
          type: string
          description: Tenant ID for multi-tenancy support
        createdAt:
          type: string
          description: Date and time when the permission was created
          format: date-time
        createdBy:
          type: string
          description: User who created the permission
        updatedAt:
          type: string
          description: Date and time when the permission was last updated
          format: date-time
        updatedBy:
          type: string
          description: User who last updated the permission
        version:
          type: integer
          description: Version number for optimistic locking
          format: int64
      description: Document permission data transfer object
...